+++ collector/src/main/java/com/sippet/collector/amqp/UserTrackSender.java
@@ -1,5 +1,6 @@
-package com.sippet.collector;
+package com.sippet.collector.amqp;
 
+import com.sippet.collector.api.UserTrackRequest;
 import com.sippet.domain.service.TrackingResult;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -18,7 +19,7 @@ public class UserTrackSender {
     @Autowired
     private UserTrackDtoConverter userTrackDtoConverter;
 
-    void send(UserTrackRequest userTrackRequest, TrackingResult trackingResult) {
+    public void send(UserTrackRequest userTrackRequest, TrackingResult trackingResult) {
         amqpTemplate.convertAndSend(userTrackDtoConverter.convert(userTrackRequest, trackingResult));
     }
 }
