+++ domain/src/main/java/com/sippet/domain/domain/usertrack/UserTrackRepository.java
@@ -1,13 +1,13 @@
 package com.sippet.domain.domain.usertrack;
 
+import com.sippet.domain.domain.usertrack.projection.UserTrackHrefCount;
+import com.sippet.domain.domain.usertrack.projection.UserTrackPathNameCount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface UserTrackRepository extends JpaRepository<UserTrack, Long> {
