+++ domain/src/main/java/com/sippet/domain/domain/usertrack/projection/UserTrackPathNameCount.java
@@ -1,4 +1,4 @@
-package com.sippet.domain.domain.usertrack;
+package com.sippet.domain.domain.usertrack.projection;
 
 public interface UserTrackPathNameCount {
     String getPathName();
